Move assignment implementation into own file
